refactor(player): extract dimension header utils to _utils folder
Move getHeaderDimensionState and getStatusColor to their own file
so they can be unit tested without @internal markers.
