+++ packages/player/src/components/dimension-header-status.tsx
@@ -12,41 +12,8 @@ import { cn } from "@zoonk/ui/lib/utils";
 import { ChevronDownIcon } from "lucide-react";
 import { useExtracted } from "next-intl";
 import { type DimensionInventory } from "../player-reducer";
+import { getHeaderDimensionState, getStatusColor } from "./_utils/dimension-header-status";
 import { DimensionList, buildDimensionEntries } from "./dimension-inventory";
-
-type HeaderDimensionState =
-  | { count: number; kind: "at-risk" }
-  | { count: number; kind: "negative" }
-  | { kind: "all-clear" };
-
-/** @internal */
-export function getHeaderDimensionState(dimensions: DimensionInventory): HeaderDimensionState {
-  const values = Object.values(dimensions);
-  const negativeCount = values.filter((value) => value < 0).length;
-
-  if (negativeCount > 0) {
-    return { count: negativeCount, kind: "negative" };
-  }
-
-  const atRiskCount = values.filter((value) => value === 0).length;
-
-  if (atRiskCount > 0) {
-    return { count: atRiskCount, kind: "at-risk" };
-  }
-
-  return { kind: "all-clear" };
-}
-
-/** @internal */
-export function getStatusColor(kind: HeaderDimensionState["kind"]): string {
-  if (kind === "negative") {
-    return "text-destructive";
-  }
-  if (kind === "at-risk") {
-    return "text-warning";
-  }
-  return "text-muted-foreground";
-}
 
 export function DimensionHeaderStatus({
   changedDimensions,
@@ -92,7 +59,7 @@ export function DimensionHeaderStatus({
         >
           {label}
         </span>
-        <ChevronDownIcon className="text-muted-foreground size-3 opacity-60" />
+        <ChevronDownIcon aria-hidden className="text-muted-foreground size-3 opacity-60" />
       </PopoverTrigger>
 
       <PopoverContent side="bottom" sideOffset={8}>
